Wrong method signatures caused problems in the high energy region with Geiss RDD. Bug fixed.

+++ include/AT_RDD.h
@@ -938,7 +938,7 @@ inline double  AT_RDD_Geiss_Gy( const double r_m,
     const double r_min_m,
     const double r_max_m,
     const double a0_m,
-    const long   norm_constant_Gy);
+    const double norm_constant_Gy);
 
 
 /**
